Go SDK: Quick adjustment to examples
